# service/task-service/api/task/v1/task_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v6.33.1
// source: task/v1/task.proto

package taskv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationTaskServiceAssignTask = "/task.v1.TaskService/AssignTask"
const OperationTaskServiceBulkAssign = "/task.v1.TaskService/BulkAssign"
const OperationTaskServiceBulkUpdateStatus = "/task.v1.TaskService/BulkUpdateStatus"
const OperationTaskServiceCreateTask = "/task.v1.TaskService/CreateTask"
const OperationTaskServiceDeleteTask = "/task.v1.TaskService/DeleteTask"
const OperationTaskServiceGetTask = "/task.v1.TaskService/GetTask"
const OperationTaskServiceGetTasksByProject = "/task.v1.TaskService/GetTasksByProject"
const OperationTaskServiceHello = "/task.v1.TaskService/Hello"
const OperationTaskServiceListTasks = "/task.v1.TaskService/ListTasks"
const OperationTaskServiceUpdateTask = "/task.v1.TaskService/UpdateTask"
const OperationTaskServiceUpdateTaskPriority = "/task.v1.TaskService/UpdateTaskPriority"
const OperationTaskServiceUpdateTaskStatus = "/task.v1.TaskService/UpdateTaskStatus"

type TaskServiceHTTPServer interface {
	AssignTask(context.Context, *AssignTaskRequest) (*UpdateTaskResponse, error)
	BulkAssign(context.Context, *BulkAssignRequest) (*BulkAssignResponse, error)
	// BulkUpdateStatus ========== BULK OPERATIONS ==========
	BulkUpdateStatus(context.Context, *BulkUpdateStatusRequest) (*BulkUpdateStatusResponse, error)
	// CreateTask ========== CREATE ==========
	CreateTask(context.Context, *CreateTaskRequest) (*CreateTaskResponse, error)
	// DeleteTask ========== DELETE ==========
	DeleteTask(context.Context, *DeleteTaskRequest) (*DeleteTaskResponse, error)
	// GetTask ========== READ ==========
	GetTask(context.Context, *GetTaskRequest) (*GetTaskResponse, error)
	GetTasksByProject(context.Context, *GetTasksByProjectRequest) (*ListTasksResponse, error)
	// Hello ========== LEGACY/HELLO (for testing) ==========
	Hello(context.Context, *HelloRequest) (*HelloResponse, error)
	ListTasks(context.Context, *ListTasksRequest) (*ListTasksResponse, error)
	// UpdateTask ========== UPDATE ==========
	UpdateTask(context.Context, *UpdateTaskRequest) (*UpdateTaskResponse, error)
	UpdateTaskPriority(context.Context, *UpdateTaskPriorityRequest) (*UpdateTaskResponse, error)
	UpdateTaskStatus(context.Context, *UpdateTaskStatusRequest) (*UpdateTaskResponse, error)
}

func RegisterTaskServiceHTTPServer(s *http.Server, srv TaskServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/tasks", _TaskService_CreateTask0_HTTP_Handler(srv))
	r.GET("/v1/tasks/{id}", _TaskService_GetTask0_HTTP_Handler(srv))
	r.GET("/v1/tasks", _TaskService_ListTasks0_HTTP_Handler(srv))
	r.GET("/v1/projects/{project_id}/tasks", _TaskService_GetTasksByProject0_HTTP_Handler(srv))
	r.PUT("/v1/tasks/{id}", _TaskService_UpdateTask0_HTTP_Handler(srv))
	r.PATCH("/v1/tasks/{id}/status", _TaskService_UpdateTaskStatus0_HTTP_Handler(srv))
	r.PATCH("/v1/tasks/{id}/assign", _TaskService_AssignTask0_HTTP_Handler(srv))
	r.PATCH("/v1/tasks/{id}/priority", _TaskService_UpdateTaskPriority0_HTTP_Handler(srv))
	r.DELETE("/v1/tasks/{id}", _TaskService_DeleteTask0_HTTP_Handler(srv))
	r.POST("/v1/tasks/bulk/status", _TaskService_BulkUpdateStatus0_HTTP_Handler(srv))
	r.POST("/v1/tasks/bulk/assign", _TaskService_BulkAssign0_HTTP_Handler(srv))
	r.GET("/v1/tasks/hello", _TaskService_Hello0_HTTP_Handler(srv))
}

func _TaskService_CreateTask0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceCreateTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateTask(ctx, req.(*CreateTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_GetTask0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTaskRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceGetTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTask(ctx, req.(*GetTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_ListTasks0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListTasksRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceListTasks)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTasks(ctx, req.(*ListTasksRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListTasksResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_GetTasksByProject0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetTasksByProjectRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceGetTasksByProject)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetTasksByProject(ctx, req.(*GetTasksByProjectRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListTasksResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_UpdateTask0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceUpdateTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTask(ctx, req.(*UpdateTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_UpdateTaskStatus0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTaskStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceUpdateTaskStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTaskStatus(ctx, req.(*UpdateTaskStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_AssignTask0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in AssignTaskRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceAssignTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.AssignTask(ctx, req.(*AssignTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_UpdateTaskPriority0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateTaskPriorityRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceUpdateTaskPriority)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateTaskPriority(ctx, req.(*UpdateTaskPriorityRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_DeleteTask0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteTaskRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceDeleteTask)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteTask(ctx, req.(*DeleteTaskRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteTaskResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_BulkUpdateStatus0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in BulkUpdateStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceBulkUpdateStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BulkUpdateStatus(ctx, req.(*BulkUpdateStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*BulkUpdateStatusResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_BulkAssign0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in BulkAssignRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceBulkAssign)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BulkAssign(ctx, req.(*BulkAssignRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*BulkAssignResponse)
		return ctx.Result(200, reply)
	}
}

func _TaskService_Hello0_HTTP_Handler(srv TaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HelloRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTaskServiceHello)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Hello(ctx, req.(*HelloRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*HelloResponse)
		return ctx.Result(200, reply)
	}
}

type TaskServiceHTTPClient interface {
	AssignTask(ctx context.Context, req *AssignTaskRequest, opts ...http.CallOption) (rsp *UpdateTaskResponse, err error)
	BulkAssign(ctx context.Context, req *BulkAssignRequest, opts ...http.CallOption) (rsp *BulkAssignResponse, err error)
	// BulkUpdateStatus ========== BULK OPERATIONS ==========
	BulkUpdateStatus(ctx context.Context, req *BulkUpdateStatusRequest, opts ...http.CallOption) (rsp *BulkUpdateStatusResponse, err error)
	// CreateTask ========== CREATE ==========
	CreateTask(ctx context.Context, req *CreateTaskRequest, opts ...http.CallOption) (rsp *CreateTaskResponse, err error)
	// DeleteTask ========== DELETE ==========
	DeleteTask(ctx context.Context, req *DeleteTaskRequest, opts ...http.CallOption) (rsp *DeleteTaskResponse, err error)
	// GetTask ========== READ ==========
	GetTask(ctx context.Context, req *GetTaskRequest, opts ...http.CallOption) (rsp *GetTaskResponse, err error)
	GetTasksByProject(ctx context.Context, req *GetTasksByProjectRequest, opts ...http.CallOption) (rsp *ListTasksResponse, err error)
	// Hello ========== LEGACY/HELLO (for testing) ==========
	Hello(ctx context.Context, req *HelloRequest, opts ...http.CallOption) (rsp *HelloResponse, err error)
	ListTasks(ctx context.Context, req *ListTasksRequest, opts ...http.CallOption) (rsp *ListTasksResponse, err error)
	// UpdateTask ========== UPDATE ==========
	UpdateTask(ctx context.Context, req *UpdateTaskRequest, opts ...http.CallOption) (rsp *UpdateTaskResponse, err error)
	UpdateTaskPriority(ctx context.Context, req *UpdateTaskPriorityRequest, opts ...http.CallOption) (rsp *UpdateTaskResponse, err error)
	UpdateTaskStatus(ctx context.Context, req *UpdateTaskStatusRequest, opts ...http.CallOption) (rsp *UpdateTaskResponse, err error)
}

type TaskServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewTaskServiceHTTPClient(client *http.Client) TaskServiceHTTPClient {
	return &TaskServiceHTTPClientImpl{client}
}

func (c *TaskServiceHTTPClientImpl) AssignTask(ctx context.Context, in *AssignTaskRequest, opts ...http.CallOption) (*UpdateTaskResponse, error) {
	var out UpdateTaskResponse
	pattern := "/v1/tasks/{id}/assign"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceAssignTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TaskServiceHTTPClientImpl) BulkAssign(ctx context.Context, in *BulkAssignRequest, opts ...http.CallOption) (*BulkAssignResponse, error) {
	var out BulkAssignResponse
	pattern := "/v1/tasks/bulk/assign"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceBulkAssign))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// BulkUpdateStatus ========== BULK OPERATIONS ==========
func (c *TaskServiceHTTPClientImpl) BulkUpdateStatus(ctx context.Context, in *BulkUpdateStatusRequest, opts ...http.CallOption) (*BulkUpdateStatusResponse, error) {
	var out BulkUpdateStatusResponse
	pattern := "/v1/tasks/bulk/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceBulkUpdateStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateTask ========== CREATE ==========
func (c *TaskServiceHTTPClientImpl) CreateTask(ctx context.Context, in *CreateTaskRequest, opts ...http.CallOption) (*CreateTaskResponse, error) {
	var out CreateTaskResponse
	pattern := "/v1/tasks"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceCreateTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteTask ========== DELETE ==========
func (c *TaskServiceHTTPClientImpl) DeleteTask(ctx context.Context, in *DeleteTaskRequest, opts ...http.CallOption) (*DeleteTaskResponse, error) {
	var out DeleteTaskResponse
	pattern := "/v1/tasks/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTaskServiceDeleteTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetTask ========== READ ==========
func (c *TaskServiceHTTPClientImpl) GetTask(ctx context.Context, in *GetTaskRequest, opts ...http.CallOption) (*GetTaskResponse, error) {
	var out GetTaskResponse
	pattern := "/v1/tasks/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTaskServiceGetTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TaskServiceHTTPClientImpl) GetTasksByProject(ctx context.Context, in *GetTasksByProjectRequest, opts ...http.CallOption) (*ListTasksResponse, error) {
	var out ListTasksResponse
	pattern := "/v1/projects/{project_id}/tasks"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTaskServiceGetTasksByProject))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Hello ========== LEGACY/HELLO (for testing) ==========
func (c *TaskServiceHTTPClientImpl) Hello(ctx context.Context, in *HelloRequest, opts ...http.CallOption) (*HelloResponse, error) {
	var out HelloResponse
	pattern := "/v1/tasks/hello"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTaskServiceHello))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TaskServiceHTTPClientImpl) ListTasks(ctx context.Context, in *ListTasksRequest, opts ...http.CallOption) (*ListTasksResponse, error) {
	var out ListTasksResponse
	pattern := "/v1/tasks"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTaskServiceListTasks))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateTask ========== UPDATE ==========
func (c *TaskServiceHTTPClientImpl) UpdateTask(ctx context.Context, in *UpdateTaskRequest, opts ...http.CallOption) (*UpdateTaskResponse, error) {
	var out UpdateTaskResponse
	pattern := "/v1/tasks/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceUpdateTask))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TaskServiceHTTPClientImpl) UpdateTaskPriority(ctx context.Context, in *UpdateTaskPriorityRequest, opts ...http.CallOption) (*UpdateTaskResponse, error) {
	var out UpdateTaskResponse
	pattern := "/v1/tasks/{id}/priority"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceUpdateTaskPriority))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *TaskServiceHTTPClientImpl) UpdateTaskStatus(ctx context.Context, in *UpdateTaskStatusRequest, opts ...http.CallOption) (*UpdateTaskResponse, error) {
	var out UpdateTaskResponse
	pattern := "/v1/tasks/{id}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationTaskServiceUpdateTaskStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PATCH", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}